+++ app/src/main/java/com/kamron/pogoiv/model/IVCombination.java
@@ -1,4 +1,4 @@
-package com.kamron.pogoiv;
+package com.kamron.pogoiv.model;
 
 /**
  * Created by Johan on 2016-08-18.
