+++ app/src/main/java/com/kamron/pogoiv/model/IVScanResult.java
@@ -1,4 +1,4 @@
-package com.kamron.pogoiv;
+package com.kamron.pogoiv.model;
 
 import java.util.ArrayList;
